remove ezvidia.aps
