+++ .gitignore
@@ -12,3 +12,4 @@
 /x64
 /EZVIDIA/EZVIDIA.vcxproj.user
 /EZVIDIA/ezconfig.json
+/EZVIDIA/ezvidia.aps
