+++ EZVIDIA/EzvidiaMaster.cpp
