+++ EZVIDIA/configurations/ConfigManager.cpp
