+++ EZVIDIA/configurations/JsonAux.cpp
