+++ EZVIDIA/include/EzvidiaMaster.hpp
