+++ EZVIDIA/include/configurations/ConfException.hpp
