+++ EZVIDIA/include/configurations/ConfigManager.hpp
