+++ EZVIDIA/include/configurations/JsonAux.hpp
