+++ EZVIDIA/include/configurations/structs/DisplayConfiguration.hpp
