+++ EZVIDIA/include/configurations/structs/GlobalConfiguration.hpp
