+++ EZVIDIA/logging.cpp
