+++ EZVIDIA/resource.h
